test: add portfolio e2e tests (#1376)
