+++ src/domains/message/cucumber/verify-message.ts
@@ -3,7 +3,7 @@ import { Selector } from "testcafe";
 import { buildTranslations } from "../../../app/i18n/helpers";
 import { cucumber, visitWelcomeScreen } from "../../../utils/e2e-utils";
 import { goToProfile } from "../../profile/e2e/common";
-import { goToWallet } from "../../wallet/e2e/common";
+import { goToWallet } from "../../portfolio/e2e/common";
 
 const translations = buildTranslations();
