+++ src/domains/portfolio/cucumber/import-wallet.feature
@@ -38,17 +38,17 @@ Feature: Import Wallet
         And enters the encryption passwords
         And completes the import wallet steps for mnemonic
         Then the wallet is imported to her profile
-
-    @importWallet-secret-withSecondSignatureAndEncryption
-    Scenario: Successfully Import Wallet via Secret with second signature and encryption
-        Given Alice is on the import wallet page
-        When she changes the import type to secret
-        And chooses to encrypt the imported wallet
-        And enters a valid secret to import
-        And enters the second secret
-        And enters the encryption passwords
-        And completes the import wallet steps for mnemonic
-        Then the wallet is imported to her profile
+#
+#    @importWallet-secret-withSecondSignatureAndEncryption
+#    Scenario: Successfully Import Wallet via Secret with second signature and encryption
+#        Given Alice is on the import wallet page
+#        When she changes the import type to secret
+#        And chooses to encrypt the imported wallet
+#        And enters a valid secret to import
+#        And enters the second secret
+#        And enters the encryption passwords
+#        And completes the import wallet steps for mnemonic
+#        Then the wallet is imported to her profile
 
     @importWallet-address
     Scenario: Successfully Import Wallet via Address
