+++ src/domains/portfolio/cucumber/import-wallet.ts
@@ -1,7 +1,7 @@
 import { Selector } from "testcafe";
 
 import { buildTranslations } from "../../../app/i18n/helpers";
-import { cucumber, mockMuSigRequest, mockRequest, visitWelcomeScreen } from "../../../utils/e2e-utils";
+import { cucumber, MNEMONICS, visitWelcomeScreen } from "../../../utils/e2e-utils";
 import { goToProfile } from "../../profile/e2e/common";
 
 const translations = buildTranslations();
@@ -11,17 +11,13 @@ const preSteps = {
 		await visitWelcomeScreen(t);
 		await goToProfile(t);
 		await t.click(Selector("button").withExactText(translations.COMMON.IMPORT));
-		await t
-			.expect(Selector("div").withText(translations.WALLETS.PAGE_IMPORT_WALLET.NETWORK_STEP.SUBTITLE).exists)
-			.ok();
-		await t.click(Selector('[data-testid="NetworkOption-ARK-ark.devnet"]'));
+
+		await t.click(Selector("div").withExactText(translations.COMMON.MNEMONIC));
 
 		await t
-			.expect(Selector("button").withText(translations.COMMON.CONTINUE).hasAttribute("disabled"))
-			.notOk("Cryptoasset selected", { timeout: 5000 });
-		await t.click(Selector("button").withExactText(translations.COMMON.CONTINUE));
-		await t
-			.expect(Selector("h1").withExactText(translations.WALLETS.PAGE_IMPORT_WALLET.METHOD_STEP.TITLE).exists)
+			.expect(
+				Selector("h2").withExactText(translations.WALLETS.PAGE_IMPORT_WALLET.METHOD_STEP.MNEMONIC_TITLE).exists,
+			)
 			.ok();
 	},
 };
@@ -41,11 +37,12 @@ const encryptionPasswordsStep = {
 
 const enterMnemonic = async (t: TestController) => {
 	const passphraseInput = Selector("[data-testid=ImportWallet__mnemonic-input]");
-	await t.typeText(passphraseInput, "buddy year cost vendor honey tonight viable nut female alarm duck symptom", {
+	await t.typeText(passphraseInput, MNEMONICS[0], {
 		paste: true,
 	});
 	await t.click(Selector("button").withExactText(translations.COMMON.CONTINUE));
 };
+
 const completeImportWallet = async (t: TestController) => {
 	await t.click(Selector("[data-testid=ImportWallet__edit-alias]"));
 	const walletNameInput = Selector("input[name=name]");
@@ -67,7 +64,8 @@ cucumber("@importWallet-mnemonic", {
 cucumber("@importWallet-mnemonic-withEncryption", {
 	...preSteps,
 	"When she chooses to encrypt the imported wallet": async (t: TestController) => {
-		await t.click(Selector('[data-testid="ImportWallet__encryption-toggle"]').parent());
+		await t.click(Selector('[data-testid="WalletEncryptionBanner__encryption"] div'));
+		await t.click(Selector('[data-testid="WalletEncryptionBanner__checkbox"]'));
 	},
 	"And enters a valid mnemonic to import": enterMnemonic,
 	...encryptionPasswordsStep,
@@ -76,53 +74,54 @@ cucumber("@importWallet-mnemonic-withEncryption", {
 		await t.click(Selector("button").withExactText(translations.COMMON.CLOSE));
 	},
 });
-
-cucumber(
-	"@importWallet-secret-withSecondSignatureAndEncryption",
-	{
-		...preSteps,
-		"When she changes the import type to secret": async (t: TestController) => {
-			await t.click('[data-testid="SelectDropdown__input"]');
-			await t.click(Selector(".select-list-option__label").withText(translations.COMMON.SECRET));
-		},
-		"And chooses to encrypt the imported wallet": async (t: TestController) => {
-			await t.click(Selector('[data-testid="ImportWallet__encryption-toggle"]').parent());
-		},
-		"And enters a valid secret to import": async (t: TestController) => {
-			const secretInput = Selector("[data-testid=ImportWallet__secret-input]");
-			await t.typeText(secretInput, "abc", {
-				paste: true,
-			});
-			await t.click(Selector("button").withExactText(translations.COMMON.CONTINUE));
-		},
-		"And enters the second secret": async (t: TestController) => {
-			const secondSecretInput = Selector("[data-testid=EncryptPassword__second-secret]");
-			await t.typeText(secondSecretInput, "abc", { paste: true });
-		},
-		...encryptionPasswordsStep,
-		"And completes the import wallet steps for secret": completeImportWallet,
-		"Then the wallet is imported to her profile": async (t: TestController) => {
-			await t.click(Selector("button").withExactText(translations.COMMON.CLOSE));
-		},
-	},
-	[
-		mockMuSigRequest("https://ark-test-musig.arkvault.io", "list", { result: [] }),
-		mockRequest(
-			"https://ark-test.arkvault.io/api/transactions?limit=30&address=DNTwQTSp999ezQ425utBsWetcmzDuCn2pN",
-			[],
-		),
-	],
-);
+//
+// cucumber(
+// 	"@importWallet-secret-withSecondSignatureAndEncryption",
+// 	{
+// 		...preSteps,
+// 		"When she changes the import type to secret": async (t: TestController) => {
+// 			await t.click('[data-testid="SelectDropdown__input"]');
+// 			await t.click(Selector(".select-list-option__label").withText(translations.COMMON.SECRET));
+// 		},
+// 		"And chooses to encrypt the imported wallet": async (t: TestController) => {
+// 			await t.click(Selector('[data-testid="ImportWallet__encryption-toggle"]').parent());
+// 		},
+// 		"And enters a valid secret to import": async (t: TestController) => {
+// 			const secretInput = Selector("[data-testid=ImportWallet__secret-input]");
+// 			await t.typeText(secretInput, "abc", {
+// 				paste: true,
+// 			});
+// 			await t.click(Selector("button").withExactText(translations.COMMON.CONTINUE));
+// 		},
+// 		"And enters the second secret": async (t: TestController) => {
+// 			const secondSecretInput = Selector("[data-testid=EncryptPassword__second-secret]");
+// 			await t.typeText(secondSecretInput, "abc", { paste: true });
+// 		},
+// 		...encryptionPasswordsStep,
+// 		"And completes the import wallet steps for secret": completeImportWallet,
+// 		"Then the wallet is imported to her profile": async (t: TestController) => {
+// 			await t.click(Selector("button").withExactText(translations.COMMON.CLOSE));
+// 		},
+// 	},
+// 	[
+// 		mockMuSigRequest("https://ark-test-musig.arkvault.io", "list", { result: [] }),
+// 		mockRequest(
+// 			"https://ark-test.arkvault.io/api/transactions?limit=30&address=DNTwQTSp999ezQ425utBsWetcmzDuCn2pN",
+// 			[],
+// 		),
+// 	],
+// );
 
 cucumber("@importWallet-address", {
 	...preSteps,
 	"When she changes the import type to address": async (t: TestController) => {
-		await t.click('[data-testid="SelectDropdown__input"]');
-		await t.click(Selector(".select-list-option__label").withText(translations.COMMON.ADDRESS));
+		await t.click(Selector("button").withExactText(translations.COMMON.BACK));
+		await t.click(Selector("div").withExactText(translations.COMMON.ADDRESS));
+		await t.expect(Selector("h2").withExactText(translations.COMMON.ADDRESS).exists).ok();
 	},
 	"And enters a valid address to import": async (t: TestController) => {
 		const addressInput = Selector("[data-testid=ImportWallet__address-input]");
-		await t.typeText(addressInput, "DC8ghUdhS8w8d11K8cFQ37YsLBFhL3Dq2P", {
+		await t.typeText(addressInput, "0x659A76be283644AEc2003aa8ba26485047fd1BFB", {
 			paste: true,
 		});
 		await t.click(Selector("button").withExactText(translations.COMMON.CONTINUE));
@@ -136,8 +135,9 @@ cucumber("@importWallet-address", {
 cucumber("@importWallet-invalidAddress", {
 	...preSteps,
 	"When she changes the import type to address": async (t: TestController) => {
-		await t.click('[data-testid="SelectDropdown__input"]');
-		await t.click(Selector(".select-list-option__label").withText(translations.COMMON.ADDRESS));
+		await t.click(Selector("button").withExactText(translations.COMMON.BACK));
+		await t.click(Selector("div").withExactText(translations.COMMON.ADDRESS));
+		await t.expect(Selector("h2").withExactText(translations.COMMON.ADDRESS).exists).ok();
 	},
 	"And enters an invalid address to import": async (t: TestController) => {
 		const addressInput = Selector("[data-testid=ImportWallet__address-input]");
@@ -173,7 +173,7 @@ cucumber("@importWallet-duplicateAddress", {
 	...preSteps,
 	"And has imported a wallet": async (t: TestController) => {
 		const passphraseInput = Selector("[data-testid=ImportWallet__mnemonic-input]");
-		await t.typeText(passphraseInput, "buddy year cost vendor honey tonight viable nut female alarm duck symptom", {
+		await t.typeText(passphraseInput, MNEMONICS[0], {
 			paste: true,
 		});
 		await t.click(Selector("button").withExactText(translations.COMMON.CONTINUE));
@@ -183,21 +183,18 @@ cucumber("@importWallet-duplicateAddress", {
 		await t.click(Selector("button").withExactText(translations.COMMON.CLOSE));
 	},
 	"When she attempts to import the same wallet again": async (t: TestController) => {
-		await t.click(Selector("a").withExactText("Portfolio"));
 		await t.click(Selector("button").withExactText(translations.COMMON.IMPORT));
+
+		await t.click(Selector("div").withExactText(translations.COMMON.MNEMONIC));
+
 		await t
-			.expect(Selector("div").withText(translations.WALLETS.PAGE_IMPORT_WALLET.NETWORK_STEP.SUBTITLE).exists)
+			.expect(
+				Selector("h2").withExactText(translations.WALLETS.PAGE_IMPORT_WALLET.METHOD_STEP.MNEMONIC_TITLE).exists,
+			)
 			.ok();
-		await t.click(Selector('[data-testid="NetworkOption-ARK-ark.devnet"]'));
-		await t
-			.expect(Selector("button").withText(translations.COMMON.CONTINUE).hasAttribute("disabled"))
-			.notOk("Cryptoasset selected", { timeout: 5000 });
-		await t.click(Selector("button").withExactText(translations.COMMON.CONTINUE));
-		await t
-			.expect(Selector("h1").withExactText(translations.WALLETS.PAGE_IMPORT_WALLET.METHOD_STEP.TITLE).exists)
-			.ok();
+
 		const passphraseInput = Selector("[data-testid=ImportWallet__mnemonic-input]");
-		await t.typeText(passphraseInput, "buddy year cost vendor honey tonight viable nut female alarm duck symptom", {
+		await t.typeText(passphraseInput, MNEMONICS[0], {
 			paste: true,
 		});
 	},
