+++ src/domains/portfolio/cucumber/portfolio-transactions.feature
@@ -1,0 +1,7 @@
+Feature: Portfolio Transactions
+
+    @portfolioTransactions-loadMore
+    Scenario: View more transactions via Load More button
+		Given Alice is signed into a profile
+        When she selects to view more transactions
+        Then the transaction count is increased
