+++ src/domains/portfolio/e2e/common.ts
